+++ nengo_loihi/hardware/interface.py
@@ -208,6 +208,7 @@ class HardwareInterface(object):
         # sort all spikes because spikegen needs them in temporal order
         loihi_spikes = sorted(loihi_spikes, key=lambda s: s.time)
         for spike in loihi_spikes:
+            assert spike.axon.axon_type == 0, "Spikegen cannot send pop spikes"
             assert spike.axon.atom == 0, "Spikegen does not support atom"
             self.n2board.global_spike_generator.addSpike(
                 time=spike.time, chipId=spike.axon.chip_id,
@@ -227,7 +228,7 @@ class HardwareInterface(object):
         assert len(loihi_spikes) <= self.snip_max_spikes_per_step
         for spike in loihi_spikes:
             assert spike.axon.chip_id == 0
-            msg.extend((spike.axon.core_id, spike.axon.axon_id))
+            msg.extend(SpikePacker.pack(spike))
         assert len(loihi_errors) == self.nengo_io_h2c_errors
         for error in loihi_errors:
             msg.extend(error)
@@ -417,15 +418,59 @@ class HardwareInterface(object):
             phase="preLearnMgmt",
         )
 
-        size = self.snip_max_spikes_per_step * 2 + 1 + total_error_len
-        logger.debug("Creating nengo_io_h2c channel")
-        self.nengo_io_h2c = self.n2board.createChannel(b'nengo_io_h2c',
-                                                       "int", size)
-        logger.debug("Creating nengo_io_c2h channel")
-        self.nengo_io_c2h = self.n2board.createChannel(b'nengo_io_c2h',
-                                                       "int", n_outputs)
+        size = (1  # first int stores number of spikes
+                + self.snip_max_spikes_per_step*SpikePacker.size()
+                + total_error_len)
+        logger.debug("Creating nengo_io_h2c channel (%d)" % size)
+        self.nengo_io_h2c = self.n2board.createChannel(
+            b'nengo_io_h2c', "int", size)
+        logger.debug("Creating nengo_io_c2h channel (%d)" % n_outputs)
+        self.nengo_io_c2h = self.n2board.createChannel(
+            b'nengo_io_c2h', "int", n_outputs)
         self.nengo_io_h2c.connect(None, nengo_io)
         self.nengo_io_c2h.connect(nengo_io, None)
         self.nengo_io_h2c_errors = n_errors
         self.nengo_io_c2h_count = n_outputs
         self.nengo_io_snip_range = snip_range
+
+
+class SpikePacker(object):
+    """Packs spikes for sending to chip
+
+    Currently represents a spike as two int32s.
+    """
+
+    @classmethod
+    def size(cls):
+        """The number of int32s used to represent one spike."""
+        size = len(cls.pack(None))
+        assert size == 2  # must match nengo_io.c.template
+        return size
+
+    @classmethod
+    def pack(cls, spike):
+        """Pack the spike into a tuple of 32-bit integers.
+
+        Parameters
+        ----------
+        spike : LoihiSpikeInput.LoihiSpike
+            The spike to pack.
+
+        Returns
+        -------
+        packed_spike : tuple of int
+            A tuple of length ``size`` to represent this spike.
+        """
+        chip_id = int(spike.axon.chip_id if spike is not None else 0)
+        core_id = int(spike.axon.core_id if spike is not None else 0)
+        axon_id = int(spike.axon.axon_id if spike is not None else 0)
+        axon_type = int(spike.axon.axon_type if spike is not None else 0)
+        atom = int(spike.axon.atom if spike is not None else 0)
+        assert chip_id == 0, "Multiple chips not supported"
+        assert 0 <= core_id < 1024
+        assert 0 <= axon_id < 4096
+        assert 0 <= axon_type <= 32
+        assert 0 <= atom < 1024
+
+        return (int(np.left_shift(core_id, 16)) + axon_id,
+                int(np.left_shift(axon_type, 16) + atom))